+++ commands/Banc/bet.js
@@ -10,7 +10,7 @@ module.exports = {
     name: 'bet',
     description: 'Aposta en un 50% de guanyar monedes amb un amic, si en tens...',
     type: 'banc',
-    cooldown: 1,
+    cooldown: 0,
     usage: '< amount/all > < @user >',
     async execute(message, args, server) {