+++ commands/Banc/slotmachine.js
@@ -11,7 +11,7 @@ module.exports = {
     name: 'slotmachine',
     description: 'Maquina tragaperras de toda la vida.\nNecessites que totes siguin iguals',
     type: 'banc',
-    cooldown: 1,
+    cooldown: 0,
     usage: '< amount/all >',
     aliases: ['slot'],
     async execute(message, args, server, userData) {