+++ commands/Entreteniment/coin.js
@@ -7,7 +7,7 @@ module.exports = {
     name: 'coin',
     description: 'Te la jugues al 50%',
     type: 'entreteniment',
-    cooldown: 1,
+    cooldown: 0,
     aliases: ['cflip', 'caraocreu'],
     execute(message) {