+++ commands/Entreteniment/howgay.js
@@ -7,7 +7,7 @@ module.exports = {
     name: 'howgay',
     description: 'Et diu lo gay que ets',
     type: 'entreteniment',
-    cooldown: 1,
+    cooldown: 0,
     aliases: ['gay'],
     execute(message, args) {