+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 _All **ÇÇ** strings are representations of the server prefix_
 
+\[v1.4\]
+
+* Afegides noves comandes de música.
+  * `ÇÇq`: Et mostra la llista de reproducció.
+  * `ÇÇnp`: Et mostra la cançó que s'està reproduint ara mateix.
+  * `ÇÇclear`: Esborra totes les cançons de la llista.
+  * Més info amb la comanda `ÇÇmusic`.
+
 \[v1.3\]
 
 * Nova comanda `lol` amb la que pots buscar tota la informació del League of Legends.